feat(auth): store tokens on Martian map
As per PR feedback, we are now storing the tokens directly on the
`Martian` map, like we do for `server-url`.

This commit also updates the `add-authentication-header` interceptor to
ensure that if an auth token is explicitly passed with the endpoint
call, we use that token, rather than the one stored in `session`. This
is useful for calling the `refresh-session` endpoint, which requires
auth with the `refresh` token, rather than the `access` token.